+++ src/blog-post/load-blog-post.js
@@ -1,8 +1,8 @@
 import marked from 'marked';
 import {BlogPost} from "./BlogPost.js";
-import {readFile} from './_deps/fs.js';
-import {parseMetadata} from './_shared/parse-metadata.js';
-import {findNextParagraphTokens, renderAbstractAsHtml} from './_shared/markdown.js';
+import {readFile} from '../_deps/fs.js';
+import {parseMetadata} from '../_shared/parse-metadata.js';
+import {findNextParagraphTokens, renderAbstractAsHtml} from '../_shared/markdown.js';
 
 const prodDeps = () => {
   return {readFile};